Refactor and fixup the existing functionality
Now, notes will properly reset everything when it runs and insert a blank note
to begin with.

Also, I've restructured the files to provide some semblance of order. Unsure
about `server/` so I kept it basic, but in `client/` it makes sense to group
everything related to one template in one folder.

Also, I like `both/` because it makes sense for an at-a-glance understanding of
the shared concerns.